send pose data via websockets
